Move chi2_shift_out function from registration task to _registration_utils.py and add registration docstring to manifest
